refactor(ingress): move the route jobs and their four job strings
`sync_instance_route`, `enqueue_route_sync`, `enqueue_route_reconcile`,
`reconcile_instance_routes` (now `reconcile`) and `_routable_instance_ips`
move from `deploy_manager` to `ingress`. `deploy_manager.py` drops to 914
lines and holds no routing.

Four dotted strings named these functions, and Frappe resolves them at run
time, so all four are rewritten in this commit: the `*/5` entry in
`hooks.py`, the method in `benchpress_settings.py`, and the two enqueue
targets that are now `benchpress.ingress.*`. The operator command in
`reconcile`'s docstring moves with it. The three RQ job ids are unchanged —
a job id is a dedup key, not an import path.

`reconcile` imports `_reconcile_bridge_attachments` inside the function
because `deploy_manager` imports this module. No re-export shim: it would
raise `ImportError` on the path an RQ worker takes.

`test_scheduler_entries.py` gains two static gates. One resolves every
method in `scheduler_events`; the other walks every non-test module for
literal `frappe.enqueue` targets and resolves those. The walker finds 15
targets today, so a later move of a job target fails a test instead of
going quiet.

Five test classes move to `test_ingress.py` with their assertions rewritten.
`TestReconcileBridgeAttachments` stays with the function it tests.
