+++ benchpress/tests/test_ingress.py
@@ -9,9 +9,12 @@ from contextlib import contextmanager
 from pathlib import Path
 from unittest.mock import MagicMock, patch
 
+import frappe
 import yaml
+from frappe.tests import IntegrationTestCase
 
 from benchpress import ingress
+from benchpress.tests.test_deploy_manager import _fresh_bench, _make_lab
 
 # Any dotted quad, anywhere in the rendered file. The property routes must hold is that no
 # address of any kind appears — asserting one known IP is absent would pass for the next one.
@@ -401,3 +404,472 @@ class TestCertificateVerification(unittest.TestCase):
 			patch.object(ingress.ssl, "create_default_context", autospec=True),
 		):
 			self.assertIsNone(ingress._certificate_error("inst-1.benchpress.cloud"))
+
+
+@contextmanager
+def _route_dir(base_domain="benchpress.cloud"):
+	"""A tmp route directory, with `base_domain` supplied rather than read from live settings.
+
+	The settings doc is patched, never saved: saving it on a real host would re-anchor the
+	certificate for whatever value a test happened to pick.
+	"""
+	with tempfile.TemporaryDirectory() as tmp:
+		target_dir = Path(tmp) / "dynamic"
+		target_dir.mkdir()
+		with (
+			patch.object(ingress, "TRAEFIK_DYNAMIC_DIR", target_dir),
+			patch("frappe.get_cached_doc", return_value=frappe._dict(base_domain=base_domain)),
+		):
+			yield ingress, target_dir
+
+
+class TestRouteConvergenceSchedule(unittest.TestCase):
+	"""The `*/5` convergence tick — see phase-3-invariant-and-convergence.md."""
+
+	def test_the_tick_hands_the_pass_to_the_long_queue(self):
+		from benchpress.ingress import enqueue_route_reconcile
+
+		with patch("frappe.enqueue") as enqueue:
+			enqueue_route_reconcile()
+
+		args, kwargs = enqueue.call_args
+		self.assertEqual(args[0], "benchpress.ingress.reconcile")
+		self.assertEqual(kwargs["queue"], "long")
+		# Fixed, so a pass running longer than the interval does not queue behind itself.
+		self.assertEqual(kwargs["job_id"], "route_reconcile")
+		self.assertTrue(kwargs["deduplicate"])
+
+	def test_the_cron_entry_is_the_enqueuer_and_never_the_pass(self):
+		"""Frappe sends cron to `default`, which `queue-short` also consumes — and that
+		container has no route mount, so the pass itself would raise every five minutes."""
+		from benchpress.hooks import scheduler_events
+
+		cron_methods = [method for methods in scheduler_events["cron"].values() for method in methods]
+
+		self.assertIn(
+			"benchpress.ingress.enqueue_route_reconcile",
+			scheduler_events["cron"]["*/5 * * * *"],
+		)
+		self.assertNotIn("benchpress.ingress.reconcile", cron_methods)
+
+
+class TestReconcileInstanceRoutes(IntegrationTestCase):
+	"""`reconcile` — the route directory converges on the database.
+
+	See specs/completed/wildcard-cert-routing/phase-3-route-convergence.md.
+	"""
+
+	@classmethod
+	def setUpClass(cls):
+		super().setUpClass()
+		frappe.set_user("Administrator")
+		cls.lab = _make_lab("test-lab-reconcile-routes")
+		frappe.db.commit()
+
+	@classmethod
+	def tearDownClass(cls):
+		frappe.set_user("Administrator")
+		for name in frappe.get_all("Bench Instance", filters={"lab": cls.lab.name}, pluck="name"):
+			frappe.delete_doc("Bench Instance", name, force=True, ignore_permissions=True)
+		cls.lab.delete(ignore_permissions=True)
+		frappe.db.commit()
+		super().tearDownClass()
+
+	def _bench(self, status, container_ip):
+		bench = _fresh_bench(self, self.lab.name)
+		bench.status = status
+		bench.container_ip = container_ip
+		bench.save(ignore_permissions=True)
+		frappe.db.commit()
+		return bench
+
+	def _seed(self, target_dir, name, body="stale\n"):
+		path = target_dir / name
+		path.write_text(body)
+		return path
+
+	def _instance_files(self, target_dir):
+		return sorted(p.name for p in target_dir.glob("*.yml") if p.name not in ingress.PROTECTED_ROUTE_FILES)
+
+	def test_a_route_file_naming_no_bench_instance_is_deleted(self):
+		"""The orphan case: a hostname with no document behind it still resolves and still
+		routes, so it keeps serving whichever container inherited that IP."""
+		with _route_dir() as (ingress, target_dir):
+			orphan = self._seed(target_dir, "091131f54bcdfc7bc37cbc45763547fa.yml")
+
+			result = ingress.reconcile()
+
+			self.assertFalse(orphan.exists())
+			self.assertGreaterEqual(result["deleted"], 1)
+
+	def test_a_bench_that_is_not_running_loses_its_route_file(self):
+		"""A stopped bench's recorded IP is an address Docker has already handed back, so the
+		file is not a dead link — it is the next bench's hostname collision."""
+		for status in ("Stopped", "Draft", "Error"):
+			with self.subTest(status=status):
+				bench = self._bench(status, "172.30.0.11")
+				with _route_dir() as (ingress, target_dir):
+					route_file = self._seed(target_dir, f"{bench.name}.yml")
+
+					ingress.reconcile()
+
+					self.assertFalse(route_file.exists())
+
+	def test_a_running_bench_route_is_rewritten_to_name_its_container(self):
+		"""Convergence, not merely reaping: a file that survives must also be right. Seeded with
+		an address backend so passing means the pass rewrote it to the container name."""
+		bench = self._bench("Running", "172.30.0.12")
+
+		with _route_dir() as (ingress, target_dir):
+			self._seed(target_dir, f"{bench.name}.yml", "http://172.30.0.99:8000\n")
+
+			result = ingress.reconcile()
+
+			written_text = (target_dir / f"{bench.name}.yml").read_text()
+			config = yaml.safe_load(written_text)
+			backends = [
+				server["url"]
+				for service in config["http"]["services"].values()
+				for server in service["loadBalancer"]["servers"]
+			]
+			self.assertIn(f"http://{bench.name}:8000", backends)
+			self.assertNotRegex(written_text, IPV4_IN_TEXT)
+			self.assertGreaterEqual(result["written"], 1)
+
+	def test_the_control_plane_router_and_the_anchor_survive_a_full_sweep(self):
+		"""The guard that stops this pass taking the platform off the internet. `dynamic.yml` is
+		the control plane's own router and the anchor is every bench's certificate; a run that
+		deletes every instance file must still leave both."""
+		with _route_dir() as (ingress, target_dir):
+			control_plane = self._seed(target_dir, "dynamic.yml", "control plane\n")
+			ingress.ensure_anchor("benchpress.cloud")
+			anchor = target_dir / "wildcard-anchor.yml"
+			anchor_text = anchor.read_text()
+			self._seed(target_dir, "16b283bccf6560ab1aa5f078d492d005.yml")
+			self._seed(target_dir, "5dc12efd9c154796adae757adec1b2f3.yml")
+
+			result = ingress.reconcile()
+
+			self.assertEqual(control_plane.read_text(), "control plane\n")
+			self.assertEqual(anchor.read_text(), anchor_text)
+			self.assertEqual(result["kept"], 2)
+
+	def test_a_run_always_leaves_the_certificate_anchored(self):
+		"""The anchor is what holds the wildcard the resolver-free bench routers serve, so the
+		pass writes it first rather than waiting for the next deploy."""
+		with _route_dir() as (ingress, target_dir):
+			result = ingress.reconcile()
+
+			self.assertTrue(result["anchored"])
+			config = yaml.safe_load((target_dir / "wildcard-anchor.yml").read_text())
+			router = config["http"]["routers"]["benchpress-wildcard-anchor"]
+			self.assertEqual(router["rule"], "Host(`tls-anchor.benchpress.cloud`)")
+
+	def test_the_returned_counts_match_what_happened_on_disk(self):
+		"""A reaper that reports what it attempted rather than what converged is how a directory
+		drifts for weeks without anyone noticing."""
+		bench = self._bench("Running", "172.30.0.13")
+
+		with _route_dir() as (ingress, target_dir):
+			self._seed(target_dir, "dynamic.yml", "control plane\n")
+			self._seed(target_dir, f"{bench.name}.yml")
+			self._seed(target_dir, "091131f54bcdfc7bc37cbc45763547fa.yml")
+			self._seed(target_dir, "5dc12efd9c154796adae757adec1b2f3.yml")
+
+			result = ingress.reconcile()
+
+			self.assertEqual(result["deleted"], 2)
+			self.assertEqual(result["kept"], 2)
+			self.assertEqual(result["written"], len(self._instance_files(target_dir)))
+			self.assertIn(f"{bench.name}.yml", self._instance_files(target_dir))
+
+	def test_a_second_run_deletes_nothing_and_touches_nothing(self):
+		"""Idempotence is the read side agreeing with the write side. A pass that keeps finding
+		things to delete is one that disagrees with the files it just wrote — and one that
+		rewrites unchanged files is a Traefik reload every five minutes, since it reloads on
+		mtime."""
+		self._bench("Running", "172.30.0.14")
+
+		with _route_dir() as (ingress, target_dir):
+			self._seed(target_dir, "091131f54bcdfc7bc37cbc45763547fa.yml")
+			ingress.reconcile()
+			before = {p.name: p.stat().st_mtime_ns for p in target_dir.iterdir()}
+
+			result = ingress.reconcile()
+
+			self.assertEqual(result["deleted"], 0)
+			self.assertFalse(result["anchored"])
+			self.assertEqual({p.name: p.stat().st_mtime_ns for p in target_dir.iterdir()}, before)
+
+	def test_reconcile_writes_nothing_at_all_without_a_public_domain(self):
+		"""A dev checkout must be byte-for-byte unaffected — and must not reap either, since a
+		directory it never writes is not a directory it understands."""
+		for base_domain in (None, "", "localhost"):
+			with (
+				self.subTest(base_domain=base_domain),
+				_route_dir(base_domain) as (
+					ingress,
+					target_dir,
+				),
+			):
+				survivor = self._seed(target_dir, "091131f54bcdfc7bc37cbc45763547fa.yml")
+
+				result = ingress.reconcile()
+
+				self.assertTrue(survivor.exists())
+				self.assertEqual(
+					result, {"anchored": False, "written": 0, "deleted": 0, "kept": 0, "attached": {}}
+				)
+
+
+class TestSettingsReAnchor(IntegrationTestCase):
+	"""`BenchPress Settings.on_update` hands the sweep to the worker that can do it.
+
+	`on_update` is called directly on an unsaved document. Saving the real Single would write
+	a domain to a live host, and the whole point of the controller is that it does not do the
+	work itself — what needs asserting is which queue it hands it to.
+	"""
+
+	def _settings(self, previous_domain):
+		settings = frappe.get_doc("BenchPress Settings")
+		settings._doc_before_save = frappe._dict(base_domain=previous_domain)
+		return settings
+
+	def test_a_changed_base_domain_enqueues_the_sweep_on_the_long_queue(self):
+		"""`queue-long` is the only worker that mounts the route directory. A sweep on any other
+		queue writes into that container's own filesystem, and Traefik never sees it."""
+		settings = self._settings("some-other-zone.example")
+
+		with patch("frappe.enqueue") as enqueue:
+			settings.on_update()
+
+		enqueue.assert_called_once()
+		args, kwargs = enqueue.call_args
+		self.assertEqual(args[0], "benchpress.ingress.reconcile")
+		self.assertEqual(kwargs["queue"], "long")
+		# The job re-reads `base_domain`, so it must not start before the new value is committed.
+		self.assertTrue(kwargs["enqueue_after_commit"])
+
+	def test_an_unrelated_settings_save_enqueues_nothing(self):
+		"""A toggle or a timeout is not a reason to sweep the route directory."""
+		settings = self._settings(frappe.get_cached_doc("BenchPress Settings").base_domain)
+
+		with patch("frappe.enqueue") as enqueue:
+			settings.on_update()
+
+		enqueue.assert_not_called()
+
+
+class TestSyncInstanceRoute(IntegrationTestCase):
+	"""`sync_instance_route` — the route directory agrees with one bench's status.
+
+	See specs/in-progress/restart-free-dynamic-routing/phase-2-status-driven-route-sync.md.
+	"""
+
+	@classmethod
+	def setUpClass(cls):
+		super().setUpClass()
+		frappe.set_user("Administrator")
+		cls.lab = _make_lab("test-lab-sync-route")
+		frappe.db.commit()
+
+	@classmethod
+	def tearDownClass(cls):
+		frappe.set_user("Administrator")
+		for name in frappe.get_all("Bench Instance", filters={"lab": cls.lab.name}, pluck="name"):
+			frappe.delete_doc("Bench Instance", name, force=True, ignore_permissions=True)
+		cls.lab.delete(ignore_permissions=True)
+		frappe.db.commit()
+		super().tearDownClass()
+
+	def _bench(self, status):
+		bench = _fresh_bench(self, self.lab.name)
+		bench.status = status
+		bench.save(ignore_permissions=True)
+		frappe.db.commit()
+		return bench
+
+	def test_a_running_bench_gets_its_route_file(self):
+		bench = self._bench("Running")
+
+		with _route_dir() as (ingress, target_dir):
+			result = ingress.sync_instance_route(bench.name)
+
+			self.assertEqual(result, "written")
+			self.assertTrue((target_dir / f"{bench.name}.yml").exists())
+
+	def test_a_stopped_bench_loses_its_route_file(self):
+		"""The live misroute closes here: a stopped bench's hostname must stop answering."""
+		bench = self._bench("Stopped")
+
+		with _route_dir() as (ingress, target_dir):
+			ingress.publish(bench.name, "benchpress.cloud")
+			route_file = target_dir / f"{bench.name}.yml"
+			self.assertTrue(route_file.exists())
+
+			result = ingress.sync_instance_route(bench.name)
+
+			self.assertEqual(result, "deleted")
+			self.assertFalse(route_file.exists())
+
+	def test_a_bench_that_no_longer_exists_is_deleted_rather_than_raised_on(self):
+		"""No status reads the same as `Stopped`. A sync that raised here would leave the
+		hostname of a deleted bench answering, which is the state this phase exists to remove."""
+		with _route_dir() as (ingress, target_dir):
+			ingress.publish("gone-for-good", "benchpress.cloud")
+
+			result = ingress.sync_instance_route("gone-for-good")
+
+			self.assertEqual(result, "deleted")
+			self.assertFalse((target_dir / "gone-for-good.yml").exists())
+
+	def test_localhost_writes_nothing_and_deletes_nothing(self):
+		"""A dev checkout has no route directory and must stay byte-for-byte unaffected."""
+		bench = self._bench("Running")
+
+		with _route_dir(base_domain="localhost") as (ingress, target_dir):
+			result = ingress.sync_instance_route(bench.name)
+
+			self.assertEqual(result, "skipped")
+			self.assertEqual(list(target_dir.iterdir()), [])
+
+	def test_localhost_leaves_an_existing_file_alone(self):
+		bench = self._bench("Stopped")
+
+		with _route_dir(base_domain="localhost") as (ingress, target_dir):
+			seeded = target_dir / f"{bench.name}.yml"
+			seeded.write_text("untouched\n")
+
+			self.assertEqual(ingress.sync_instance_route(bench.name), "skipped")
+			self.assertEqual(seeded.read_text(), "untouched\n")
+
+	def test_the_written_route_names_the_container_and_no_address(self):
+		"""Phase 1's property has to survive this path too, since it is now the common one."""
+		bench = self._bench("Running")
+
+		with _route_dir() as (ingress, target_dir):
+			ingress.sync_instance_route(bench.name)
+			written = (target_dir / f"{bench.name}.yml").read_text()
+
+		self.assertIn(f"http://{bench.name}:8000", written)
+		self.assertNotRegex(written, IPV4_IN_TEXT)
+
+
+class TestRouteSyncTriggers(IntegrationTestCase):
+	"""Every status transition hands the route write to `queue-long`.
+
+	`queue-long` is the only worker that mounts the route directory. `backend` serves these
+	requests and mounts it not at all, and `default` is consumed by `queue-short`, which has no
+	mount either — so a sync on any other queue writes into that container's own filesystem and
+	Traefik never sees it, with every check green. The queue argument is the property under test.
+	"""
+
+	@classmethod
+	def setUpClass(cls):
+		super().setUpClass()
+		frappe.set_user("Administrator")
+		cls.lab = _make_lab("test-lab-route-sync-triggers")
+		frappe.db.commit()
+
+	@classmethod
+	def tearDownClass(cls):
+		frappe.set_user("Administrator")
+		for name in frappe.get_all("Bench Instance", filters={"lab": cls.lab.name}, pluck="name"):
+			frappe.delete_doc("Bench Instance", name, force=True, ignore_permissions=True)
+		cls.lab.delete(ignore_permissions=True)
+		frappe.db.commit()
+		super().tearDownClass()
+
+	def _bench(self, status="Running"):
+		bench = _fresh_bench(self, self.lab.name)
+		bench.status = status
+		bench.container_id = "container-route-sync"
+		bench.save(ignore_permissions=True)
+		frappe.db.commit()
+		return bench
+
+	def _assert_synced_on_long(self, enqueue, bench_name):
+		enqueue.assert_called_once()
+		args, kwargs = enqueue.call_args
+		self.assertEqual(args[0], "benchpress.ingress.sync_instance_route")
+		self.assertEqual(kwargs["bench_name"], bench_name)
+		self.assertEqual(kwargs["queue"], "long")
+		# The job re-reads `status`, so it must not start before the new value is committed.
+		self.assertTrue(kwargs["enqueue_after_commit"])
+
+	@patch("benchpress.deploy_manager.stop_container")
+	def test_stop_enqueues_the_sync_on_the_long_queue(self, mock_stop):
+		from benchpress.deploy_manager import stop_bench
+
+		bench = self._bench()
+
+		with patch("frappe.enqueue") as enqueue:
+			stop_bench(bench.name)
+
+		self._assert_synced_on_long(enqueue, bench.name)
+
+	def test_start_enqueues_the_sync_on_the_long_queue(self):
+		from benchpress.api import bench_action
+
+		bench = self._bench("Stopped")
+
+		with patch("benchpress.docker_manager.start_container"), patch("frappe.enqueue") as enqueue:
+			bench_action(bench.name, "start")
+
+		self._assert_synced_on_long(enqueue, bench.name)
+
+	def test_restart_enqueues_the_sync_on_the_long_queue(self):
+		"""`docker restart` re-allocates the address, and the file may have been removed by a
+		previous stop — a restart that skipped the sync would come back unrouted."""
+		from benchpress.api import bench_action
+
+		bench = self._bench()
+
+		with patch("benchpress.docker_manager.restart_container"), patch("frappe.enqueue") as enqueue:
+			bench_action(bench.name, "restart")
+
+		self._assert_synced_on_long(enqueue, bench.name)
+
+	def test_desk_start_enqueues_the_sync_on_the_long_queue(self):
+		bench = self._bench("Stopped")
+
+		with (
+			patch("benchpress.docker_manager.start_container"),
+			patch("frappe.msgprint"),
+			patch("frappe.enqueue") as enqueue,
+		):
+			bench.enqueue_start()
+
+		self._assert_synced_on_long(enqueue, bench.name)
+
+	def test_two_transitions_in_quick_succession_deduplicate(self):
+		"""A stop followed straight away by a start must not queue two writes racing each other
+		for the same file — the job id is per bench and `deduplicate` collapses them."""
+		from benchpress.ingress import enqueue_route_sync
+
+		with patch("frappe.enqueue") as enqueue:
+			enqueue_route_sync("inst-1")
+			enqueue_route_sync("inst-1")
+
+		job_ids = {call.kwargs["job_id"] for call in enqueue.call_args_list}
+		self.assertEqual(job_ids, {"route_sync:inst-1"})
+		self.assertTrue(all(call.kwargs["deduplicate"] for call in enqueue.call_args_list))
+
+	def test_teardown_deletes_directly_and_enqueues_nothing(self):
+		"""Teardown already runs on `queue-long`, and it must not depend on a second job
+		surviving to remove live routing state."""
+		from benchpress.deploy_manager import teardown_bench
+
+		bench = self._bench()
+		bench.container_id = None
+		bench.save(ignore_permissions=True)
+
+		with _route_dir() as (_ingress, target_dir):
+			ingress.publish(bench.name, "benchpress.cloud")
+			route_file = target_dir / f"{bench.name}.yml"
+
+			with patch("frappe.enqueue") as enqueue:
+				teardown_bench(bench)
+
+			self.assertFalse(route_file.exists())
+			enqueue.assert_not_called()
